Move Storage.java + new folder structure + refactorings
- Main class and resources package moved to server package.
- json2Build function moved from Resource.java to Build.java, implemented as constructor.
- Storage object is initialized in CIServer class as static variable, other classes using storage should import it from there.
- Minor warnings fixed in BuildJob.java
- Unused imports removed
