+++ build.gradle
@@ -27,5 +27,5 @@ dependencies {
 }
 
 application {
-    mainClassName = 'ContinuousIntegrationServer'
+    mainClassName = 'server.ContinuousIntegrationServer'
 }
