+++ src/main/java/buildtools/Storage.java
@@ -1,23 +1,18 @@
-/**
- * Stores and fetches data from a json file regarding build info
- */
-package resources;
-
-import buildtools.Build;
+package buildtools;
 
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.File;
 
-import org.eclipse.jetty.websocket.api.WebSocketBehavior;
-import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.JSONException;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
+/**
+ * Stores and fetches data from a json file regarding build info
+ */
 public class Storage {
     String fileName = "db.json";
     /**
@@ -43,8 +38,8 @@ public class Storage {
      * Taken from https://stackoverflow.com/questions/7463414/what-s-the-best-way-to-load-a-jsonobject-from-a-json-text-file
      * @param file name of file name
      * @return json object of the file given
-     * @throws JSONException
-     * @throws IOException
+     * @throws JSONException - file content is not valid JSON.
+     * @throws IOException - cannot read from file with name @param file
      */
     private JSONObject parseJSON(String file) throws JSONException, IOException {
         String content = new String(Files.readAllBytes(Paths.get(file)));
@@ -52,11 +47,9 @@ public class Storage {
     }
 
     /**
-     * Adds a new commit to the data object
-     * @param id id of the commit
-     * @param passed checks passed
-     * @param header commit header
-     * @param body commit message
+     * Writes a build to the database file in JSON format.
+     * @param build - the Build object to be stored.
+     * @throws IOException - if db file is not found.
      */
     public void post(Build build) throws IOException {
         File db = new File(fileName);
@@ -82,9 +75,10 @@ public class Storage {
     }
 
     /**
-     * Fetches a specific commit from the data object.
-     * @param id id of commit to get
-     * @return object with given id or empty object if not found
+     * Reads a build from the database.
+     * @param jobID - the build's jobID.
+     * @return build as Build object
+     * @throws IOException - if db file is not found
      */
     public JSONObject get(String jobID) throws IOException {
         File db = new File(fileName);
@@ -106,8 +100,8 @@ public class Storage {
     }
 
     /**
-     * Gets all the commits in the database
-     * @return all commits as a JSONObject
+     * Gets all build from the database.
+     * @return all builds as a JSONObject.
      * @throws IOException if database file was not found
      */
     public JSONObject getAll() throws IOException {
@@ -115,9 +109,7 @@ public class Storage {
         if (!db.exists()) {
             throw new IOException("Database file not found");
         } else {
-            JSONObject dbJSON = parseJSON(fileName);
-            
-            return dbJSON;
+            return parseJSON(fileName);
         }
     }
 }
