+++ src/main/java/server/ContinuousIntegrationServer.java
@@ -1,11 +1,8 @@
+package server;
+
+import buildtools.Storage;
 import org.eclipse.jetty.server.Server;
 
-import java.util.*;
-
-/** 
- Skeleton of a ContinuousIntegrationServer which acts as webhook
- See the Jetty documentation for API documentation of those classes.
-*/
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
 import org.glassfish.jersey.servlet.ServletContainer;
@@ -14,12 +11,17 @@ import org.slf4j.LoggerFactory;
 
 import static org.eclipse.jetty.servlet.ServletContextHandler.NO_SESSIONS;
 
+/**
+ Skeleton of a server.ContinuousIntegrationServer which acts as webhook
+ See the Jetty documentation for API documentation of those classes.
+ */
 public class ContinuousIntegrationServer {
+
 
     // prints output, errors and exit value of bash executions
 
-
     private static final Logger logger = LoggerFactory.getLogger(ContinuousIntegrationServer.class);
+    public static Storage storage = new Storage();
 
     public static void main(String[] args) {
 
@@ -32,7 +34,7 @@ public class ContinuousIntegrationServer {
 
         ServletHolder servletHolder = servletContextHandler.addServlet(ServletContainer.class, "/*");
         servletHolder.setInitOrder(0);
-        servletHolder.setInitParameter("jersey.config.server.provider.packages", "resources");
+        servletHolder.setInitParameter("jersey.config.server.provider.packages", "server/resources");
 
 
         //printBash(commands);
